@@ -12,40 +12,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/cdk": "^5.2.1",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.1",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "angular2-jwt": "^0.2.3",
-    "angular2-materialize": "^15.1.10",
-    "angular2-multiselect-dropdown": "^2.5.0",
+    "@angular/animations": "~8.2.9",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "~8.2.9",
+    "@angular/compiler": "~8.2.9",
+    "@angular/core": "~8.2.9",
+    "@angular/forms": "~8.2.9",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.9",
+    "@angular/platform-browser-dynamic": "~8.2.9",
+    "@angular/router": "~8.2.9",
+    "@auth0/angular-jwt": "^3.0.0",
     "bootstrap": "^4.0.0",
     "core-js": "^2.4.1",
     "jquery": "^2.2.4",
     "materialize-css": "^0.100.2",
     "mdi": "^2.1.19",
     "moment": "^2.20.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.0.0-rc.0",
     "socket.io": "^2.0.4",
     "socket.io-client": "^2.0.4",
     "socketio-wildcard": "^2.0.0",
-    "zone.js": "^0.8.19"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.7",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.803.9",
+    "@angular/cli": "^8.3.9",
+    "@angular/compiler-cli": "~8.2.9",
+    "@angular/language-service": "~8.2.9",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -57,6 +57,7 @@
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
     "tslint-eslint-rules": "^5.4.0",
-    "typescript": "~2.5.3"
+    "typescript": "~3.5.3",
+    "webpack": "^4.41.2"
   }
 }
